# api/openapi.yaml
openapi: 3.1.0
info:
  title: 神州 HR 考勤系统 API
  version: 1.9.0-wave3
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    V1.9 WAVE-3 契约：在完整保留 WAVE-1 本地账号、登录会话、权限、审计与
    版本化规则配置底座，以及 WAVE-2 组织、员工期初导入与本地人员权威的基础上，
    增加考勤地点、考勤组、员工期限分配、不可变班次版本、工作日历和基础策略绑定。
    所有 ID 均按 string 传输；响应和示例不包含密码、密码哈希、session token 或 reset token。
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /access/accounts:
    get:
      tags: [Access]
      operationId: listAccounts
      summary: 分页查询账号
      x-capability: ACCOUNT:READ
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
        - name: query
          in: query
          schema: { type: string, maxLength: 100 }
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AccountStatus'
      responses:
        '200':
          description: 授权范围内账号
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountPage'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
    post:
      tags: [Access]
      operationId: createAccount
      summary: 创建本地账号
      x-capability: ACCOUNT:CREATE
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreateRequest'
      responses:
        '201':
          description: 账号已创建；响应不返回初始密码
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDetail'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '409':
          $ref: '#/components/responses/Conflict'
  /access/accounts/{accountId}:
    get:
      tags: [Access]
      operationId: getAccount
      summary: 获取账号详情、角色和会话摘要
      x-capability: ACCOUNT:READ
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: 账号详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDetail'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
  /access/accounts/{accountId}/status:
    patch:
      tags: [Access]
      operationId: updateAccountStatus
      summary: 启用或停用账号；停用会撤销现有会话
      x-capability: ACCOUNT:EDIT
      parameters:
        - $ref: '#/components/parameters/AccountId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountStatusUpdateRequest'
      responses:
        '200':
          description: 更新后的账号
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDetail'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
        '409':
          $ref: '#/components/responses/StaleVersion'
  /access/accounts/{accountId}/lock:
    post:
      tags: [Access]
      operationId: lockAccount
      summary: 管理员锁定账号并撤销会话
      x-capability: ACCOUNT:LOCK
      parameters:
        - $ref: '#/components/parameters/AccountId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
      responses:
        '204':
          description: 已锁定
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
  /access/accounts/{accountId}/unlock:
    post:
      tags: [Access]
      operationId: unlockAccount
      summary: 管理员解除账号锁定
      x-capability: ACCOUNT:UNLOCK
      parameters:
        - $ref: '#/components/parameters/AccountId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
      responses:
        '204':
          description: 已解锁
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
  /access/accounts/{accountId}/password-reset-grants:
    post:
      tags: [Access]
      operationId: issuePasswordResetGrant
      summary: 管理员发起密码重置授权
      x-capability: ACCOUNT:RESET_PASSWORD
      parameters:
        - $ref: '#/components/parameters/AccountId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
      responses:
        '202':
          description: 已建立短时一次性授权并通过仓库外安全渠道交付
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedOperation'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
  /access/accounts/{accountId}/role-assignments:
    put:
      tags: [Access]
      operationId: replaceAccountRoleAssignments
      summary: 按有效期和数据范围替换账号角色授权
      x-capability: ROLE:ASSIGN
      parameters:
        - $ref: '#/components/parameters/AccountId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleAssignmentsUpdateRequest'
      responses:
        '200':
          description: 更新后的账号详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDetail'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/ResourceNotAvailable'
        '409':
          $ref: '#/components/responses/StaleVersion'
  /access/roles:
    get:
      tags: [Access]
      operationId: listRoles
      summary: 查询角色、能力和数据范围定义
      x-capability: ROLE:READ
      responses:
        '200':
          description: 角色列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleView'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SHENZHOUHR_SESSION
      description: 本系统本地账号安全会话；开发身份头不属于正式认证路径。
...
  parameters:
    CsrfToken:
      name: X-CSRF-TOKEN
      in: header
      required: true
      schema: { type: string, minLength: 32, maxLength: 256 }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Sort:
      name: sort
      in: query
      schema:
        type: string
        enum: [createdAt,updatedAt,username,status,occurredAt,templateCode,versionNumber]
    AccountId:
      name: accountId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ResourceId' }
...
  schemas:
    ResourceId:
      type: string
      minLength: 1
      maxLength: 128
      description: 精确资源标识；禁止转换为 JavaScript Number。
...
    AccountStatus:
      type: string
      enum: [ACTIVE, DISABLED, LOCKED]
    SessionStatus:
      type: string
      enum: [ACTIVE, EXPIRED, REVOKED]
...
    AcceptedOperation:
      type: object
      additionalProperties: false
      required: [accepted, correlationId]
      properties:
        accepted: { type: boolean, const: true }
        correlationId: { type: string }
    ReasonRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason: { type: string, minLength: 2, maxLength: 500 }
    AccountSummary:
      type: object
      additionalProperties: false
      required: [accountId, username, displayName, status, firstPasswordChangeRequired, rowVersion]
      properties:
        accountId: { $ref: '#/components/schemas/ResourceId' }
        username: { type: string }
        displayName: { type: string }
        status: { $ref: '#/components/schemas/AccountStatus' }
        firstPasswordChangeRequired: { type: boolean }
        resetPending: { type: boolean }
        lockedUntil: { type: [string, 'null'], format: date-time }
        lastLoginAt: { type: [string, 'null'], format: date-time }
        rowVersion: { type: integer, format: int64, minimum: 0 }
    AccountDetail:
      allOf:
        - $ref: '#/components/schemas/AccountSummary'
        - type: object
          required: [roles, sessions]
          properties:
            roles:
              type: array
              items:
                $ref: '#/components/schemas/RoleAssignmentView'
            sessions:
              type: array
              items:
                $ref: '#/components/schemas/SessionSummary'
    AccountPage:
      type: object
      additionalProperties: false
      required: [items, total, page, size]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AccountSummary'
        total: { type: integer, format: int64, minimum: 0 }
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1, maximum: 100 }
    AccountCreateRequest:
      type: object
      additionalProperties: false
      required: [username, displayName, temporaryPassword, roleAssignments]
      properties:
        username: { type: string, minLength: 3, maxLength: 128 }
        displayName: { type: string, minLength: 1, maxLength: 100 }
        temporaryPassword: { type: string, minLength: 12, maxLength: 256, writeOnly: true }
        roleAssignments:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/RoleAssignmentRequest'
    AccountStatusUpdateRequest:
      type: object
      additionalProperties: false
      required: [status, reason, expectedVersion]
      properties:
        status: { type: string, enum: [ACTIVE, DISABLED] }
        reason: { type: string, minLength: 2, maxLength: 500 }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
    SessionSummary:
      type: object
      additionalProperties: false
      required: [sessionId, status, createdAt, idleExpiresAt, absoluteExpiresAt]
      properties:
        sessionId: { $ref: '#/components/schemas/ResourceId' }
        status: { $ref: '#/components/schemas/SessionStatus' }
        createdAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time }
        idleExpiresAt: { type: string, format: date-time }
        absoluteExpiresAt: { type: string, format: date-time }
        revokedAt: { type: [string, 'null'], format: date-time }
        revocationReason: { type: [string, 'null'] }
    RoleView:
      type: object
      additionalProperties: false
      required: [roleId, code, name, capabilities]
      properties:
        roleId: { $ref: '#/components/schemas/ResourceId' }
        code: { type: string }
        name: { type: string }
        capabilities:
          type: array
          uniqueItems: true
          items: { type: string }
    RoleAssignmentRequest:
      type: object
      additionalProperties: false
      required: [roleId, scopeType, validFrom]
      properties:
        roleId: { $ref: '#/components/schemas/ResourceId' }
        scopeType: { type: string, enum: [LEGAL_ENTITY, ORGANIZATION, SELF] }
        scopeResourceId:
          oneOf:
            - $ref: '#/components/schemas/ResourceId'
            - type: 'null'
        validFrom: { type: string, format: date-time }
        validTo: { type: [string, 'null'], format: date-time }
    RoleAssignmentView:
      allOf:
        - $ref: '#/components/schemas/RoleAssignmentRequest'
        - type: object
          required: [assignmentId, roleCode, roleName]
          properties:
            assignmentId: { $ref: '#/components/schemas/ResourceId' }
            roleCode: { type: string }
            roleName: { type: string }
    RoleAssignmentsUpdateRequest:
      type: object
      additionalProperties: false
      required: [assignments, reason, expectedVersion]
      properties:
        assignments:
          type: array
          items:
            $ref: '#/components/schemas/RoleAssignmentRequest'
        reason: { type: string, minLength: 2, maxLength: 500 }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
...
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, correlationId, retryable]
      properties:
        code: { type: string }
        message: { type: string }
        correlationId: { type: string }
        retryable: { type: boolean }
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        details:
          type: object
          additionalProperties: true
  responses:
    AuthenticationRequired:
      description: 身份无效、会话到期或会话已撤销
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    AccessDenied:
      description: 已认证但动作、数据范围、记录状态或会话条件禁止
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ResourceNotAvailable:
      description: 对象不存在或主体无权读取，响应不区分原因
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ValidationFailed:
      description: 请求字段、受控枚举或业务校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: 状态、范围、生效期、冻结保护或唯一性冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    StaleVersion:
      description: 乐观锁版本过期
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...